Polish xp3 handling a bit more
It's really not perfect, I wish it was, but XP3 is hard man D:

+++ src/ArchiveFormats/Entry.h
@@ -20,7 +20,7 @@ struct Entry {
     uint64_t size;
     time_t lastModified;
 
-    uint32_t packedSize;
+    uint64_t packedSize;
     bool isPacked;
     uint64_t index;
